backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
aiohttp>=3.9.0
python-dotenv==0.19.2
pytz==2021.3
schedule==1.1.0
twilio==7.8.0
google-genai